Mine/requirements.txt
kiteconnect==3.1.0
python-dotenv==1.0.0
flask==3.0.0
pandas==2.1.0
pytest==7.4.0
black
flake8